Changed long to size_t as datatype of the hash.

+++ support/InstROMeasurementInterface.c
@@ -1,26 +1,26 @@
 #include "InstROMeasurementInterface.h"
 
 
-void __instro_start_function(long construct_hash, void *callsite) {}
+void __instro_start_function(size_t construct_hash, void *callsite) {}
 
-void __instro_end_function(long construct_hash, void *callsite) {}
+void __instro_end_function(size_t construct_hash, void *callsite) {}
 
-void __instro_start_statement(long construct_hash) {}
+void __instro_start_statement(size_t construct_hash) {}
 
-void __instro_end_statement(long construct_hash) {}
+void __instro_end_statement(size_t construct_hash) {}
 
-void __instro_start_loop(long construct_hash) {}
+void __instro_start_loop(size_t construct_hash) {}
 
-void __instro_end_loop(long construct_hash) {}
+void __instro_end_loop(size_t construct_hash) {}
 
-void __instro_start_scope(long construct_hash) {}
+void __instro_start_scope(size_t construct_hash) {}
 
-void __instro_end_scope(long construct_hash) {}
+void __instro_end_scope(size_t construct_hash) {}
 
-void __instro_start_conditional(long construct_hash) {}
+void __instro_start_conditional(size_t construct_hash) {}
 
-void __instro_end_conditional(long construct_hash) {}
+void __instro_end_conditional(size_t construct_hash) {}
 
-void __instro_start_expression(long construct_hash) {}
+void __instro_start_expression(size_t construct_hash) {}
 
-void __instro_end_expression(long construct_hash) {}
+void __instro_end_expression(size_t construct_hash) {}
